refactor: separate static functions from sshnpd, offload cli atClient Creation and parameterization
